drm/nouveau/nvif: use negative oclass identifier for internal classes

+++ drivers/gpu/drm/nouveau/nvkm/core/parent.c
@@ -27,7 +27,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #include <core/engine.h>
 
 int
-nvkm_parent_sclass(struct nvkm_object *parent, u16 handle,
+nvkm_parent_sclass(struct nvkm_object *parent, s32 handle,
 		   struct nvkm_object **pengine,
 		   struct nvkm_oclass **poclass)
 {
@@ -67,7 +67,7 @@ nvkm_parent_sclass(struct nvkm_object *parent, u16 handle,
 }
 
 int
-nvkm_parent_lclass(struct nvkm_object *parent, u32 *lclass, int size)
+nvkm_parent_lclass(struct nvkm_object *parent, s32 *lclass, int size)
 {
 	struct nvkm_oclass *sclass, *oclass;
 	struct nvkm_engine *engine;
